Create an ObjectInspectorItem component
