+++ packages/devtools-reps/src/object-inspector/components/ObjectInspector.js
@@ -5,52 +5,30 @@
 // @flow
 
 const { Component, createFactory, createElement } = require("react");
-const dom = require("react-dom-factories");
 const { connect } = require("react-redux");
 const actions = require("../actions");
 
 const selectors = require("../reducer");
 
-import Services from "devtools-services";
-const { appinfo } = Services;
-const isMacOS = appinfo.OS === "Darwin";
-
 import Components from "devtools-components";
 const Tree = createFactory(Components.Tree);
 require("./ObjectInspector.css");
 
+const ObjectInspectorItem = createFactory(require("./ObjectInspectorItem"));
+
 const classnames = require("classnames");
-const { MODE } = require("../../reps/constants");
 
 const Utils = require("../utils");
 const { renderRep, shouldRenderRootsInReps } = Utils;
-
 const {
   getChildren,
   getActor,
   getParent,
-  getValue,
   nodeHasAccessors,
-  nodeHasProperties,
-  nodeIsBlock,
-  nodeIsDefaultProperties,
-  nodeIsFunction,
-  nodeIsGetter,
-  nodeIsMapEntry,
-  nodeIsMissingArguments,
-  nodeIsOptimizedOut,
-  nodeIsPrimitive,
-  nodeIsPrototype,
-  nodeIsSetter,
-  nodeIsUninitializedBinding,
-  nodeIsUnmappedBinding,
-  nodeIsUnscopedBinding,
-  nodeIsWindow,
-  nodeIsLongString,
-  nodeHasFullText
+  nodeIsPrimitive
 } = Utils.node;
 
-import type { CachedNodes, Node, NodeContents, Props } from "../types";
+import type { CachedNodes, Props } from "../types";
 
 type DefaultProps = {
   autoExpandAll: boolean,
@@ -93,7 +71,6 @@ class ObjectInspector extends Component<Props> {
     const self: any = this;
 
     self.getItemChildren = this.getItemChildren.bind(this);
-    self.renderTreeItem = this.renderTreeItem.bind(this);
     self.setExpanded = this.setExpanded.bind(this);
     self.focusItem = this.focusItem.bind(this);
     self.getRoots = this.getRoots.bind(this);
@@ -206,241 +183,6 @@ class ObjectInspector extends Component<Props> {
     }
   }
 
-  // eslint-disable-next-line complexity
-  getTreeItemLabelAndValue(
-    item: Node,
-    depth: number,
-    expanded: boolean
-  ): {
-    value?: string | Element,
-    label?: string
-  } {
-    const label = item.name;
-    const isPrimitive = nodeIsPrimitive(item);
-
-    if (nodeIsOptimizedOut(item)) {
-      return {
-        label,
-        value: dom.span({ className: "unavailable" }, "(optimized away)")
-      };
-    }
-
-    if (nodeIsUninitializedBinding(item)) {
-      return {
-        label,
-        value: dom.span({ className: "unavailable" }, "(uninitialized)")
-      };
-    }
-
-    if (nodeIsUnmappedBinding(item)) {
-      return {
-        label,
-        value: dom.span({ className: "unavailable" }, "(unmapped)")
-      };
-    }
-
-    if (nodeIsUnscopedBinding(item)) {
-      return {
-        label,
-        value: dom.span({ className: "unavailable" }, "(unscoped)")
-      };
-    }
-
-    const itemValue = getValue(item);
-    const unavailable =
-      isPrimitive &&
-      itemValue &&
-      itemValue.hasOwnProperty &&
-      itemValue.hasOwnProperty("unavailable");
-
-    if (nodeIsMissingArguments(item) || unavailable) {
-      return {
-        label,
-        value: dom.span({ className: "unavailable" }, "(unavailable)")
-      };
-    }
-
-    if (
-      nodeIsFunction(item) &&
-      !nodeIsGetter(item) &&
-      !nodeIsSetter(item) &&
-      (this.props.mode === MODE.TINY || !this.props.mode)
-    ) {
-      return {
-        label: Utils.renderRep(item, {
-          ...this.props,
-          functionName: label
-        })
-      };
-    }
-
-    if (
-      nodeHasProperties(item) ||
-      nodeHasAccessors(item) ||
-      nodeIsMapEntry(item) ||
-      nodeIsLongString(item) ||
-      isPrimitive
-    ) {
-      const repProps = { ...this.props };
-      if (depth > 0) {
-        repProps.mode = this.props.mode === MODE.LONG ? MODE.SHORT : MODE.TINY;
-      }
-      if (expanded) {
-        repProps.mode = MODE.TINY;
-      }
-
-      if (nodeIsLongString(item)) {
-        repProps.member = {
-          open: nodeHasFullText(item) && expanded
-        };
-      }
-
-      return {
-        label,
-        value: Utils.renderRep(item, repProps)
-      };
-    }
-
-    return {
-      label
-    };
-  }
-
-  renderTreeItemLabel(
-    label,
-    item: Node,
-    depth: number,
-    focused: boolean,
-    expanded: boolean
-  ) {
-    if (label === null || typeof label === "undefined") {
-      return null;
-    }
-
-    const { onLabelClick } = this.props;
-
-    return dom.span(
-      {
-        className: "object-label",
-        onClick: onLabelClick
-          ? event => {
-              event.stopPropagation();
-
-              // If the user selected text, bail out.
-              if (Utils.selection.documentHasSelection()) {
-                return;
-              }
-
-              onLabelClick(item, {
-                depth,
-                focused,
-                expanded,
-                setExpanded: this.setExpanded
-              });
-            }
-          : undefined
-      },
-      label
-    );
-  }
-
-  getTreeTopElementProps(
-    item: Node,
-    depth: number,
-    focused: boolean,
-    expanded: boolean
-  ): Object {
-    const { onCmdCtrlClick, onDoubleClick, dimTopLevelWindow } = this.props;
-
-    const parentElementProps: Object = {
-      className: classnames("node object-node", {
-        focused,
-        lessen:
-          !expanded &&
-          (nodeIsDefaultProperties(item) ||
-            nodeIsPrototype(item) ||
-            nodeIsGetter(item) ||
-            nodeIsSetter(item) ||
-            (dimTopLevelWindow === true && nodeIsWindow(item) && depth === 0)),
-        block: nodeIsBlock(item)
-      }),
-      onClick: e => {
-        if (
-          onCmdCtrlClick &&
-          ((isMacOS && e.metaKey) || (!isMacOS && e.ctrlKey))
-        ) {
-          onCmdCtrlClick(item, {
-            depth,
-            event: e,
-            focused,
-            expanded
-          });
-          e.stopPropagation();
-          return;
-        }
-
-        // If this click happened because the user selected some text, bail out.
-        // Note that if the user selected some text before and then clicks here,
-        // the previously selected text will be first unselected, unless the
-        // user clicked on the arrow itself. Indeed because the arrow is an
-        // image, clicking on it does not remove any existing text selection.
-        // So we need to also check if the arrow was clicked.
-        if (
-          Utils.selection.documentHasSelection() &&
-          !(e.target && e.target.matches && e.target.matches(".arrow"))
-        ) {
-          e.stopPropagation();
-        }
-      }
-    };
-
-    if (onDoubleClick) {
-      parentElementProps.onDoubleClick = e => {
-        e.stopPropagation();
-        onDoubleClick(item, {
-          depth,
-          focused,
-          expanded
-        });
-      };
-    }
-
-    return parentElementProps;
-  }
-
-  renderTreeItem(
-    item: Node,
-    depth: number,
-    focused: boolean,
-    arrow: Object,
-    expanded: boolean
-  ) {
-    const { label, value } = this.getTreeItemLabelAndValue(
-      item,
-      depth,
-      expanded
-    );
-    const labelElement = this.renderTreeItemLabel(
-      label,
-      item,
-      depth,
-      focused,
-      expanded
-    );
-    const delimiter =
-      value && labelElement
-        ? dom.span({ className: "object-delimiter" }, ": ")
-        : null;
-
-    return dom.div(
-      this.getTreeTopElementProps(item, depth, focused, expanded),
-      arrow,
-      labelElement,
-      delimiter,
-      value
-    );
-  }
-
   render() {
     const {
       autoExpandAll = true,
@@ -476,14 +218,22 @@ class ObjectInspector extends Component<Props> {
       onCollapse: item => this.setExpanded(item, false),
       onFocus: focusable ? this.focusItem : null,
 
-      renderItem: this.renderTreeItem
+      renderItem: (item, depth, focused, arrow, expanded) =>
+        ObjectInspectorItem({
+          item,
+          depth,
+          focused,
+          arrow,
+          expanded,
+          ...this.props,
+          setExpanded: this.setExpanded
+        })
     });
   }
 }
 
 function mapStateToProps(state, props) {
   return {
-    actors: selectors.getActors(state),
     expandedPaths: selectors.getExpandedPaths(state),
     loadedProperties: selectors.getLoadedProperties(state)
   };
